feat: Added 404 route and updated ingredient quantity

+++ backend/src/models.rs
@@ -78,7 +78,7 @@ pub struct Ingredient {
     pub category: String,
 
     #[serde(rename = "Quantity")]
-    pub quantity: u16,
+    pub quantity: f32,
 
     #[serde(rename = "Unit")]
     pub unit: String,
